API access token

+++ src/pages/api/[[...user]].ts
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "GET") {
@@ -15,21 +16,38 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       data,
     });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData("users", id, data, (result: boolean) => {
-      if (result) {
-        res.status(200).json({ status: true, statusCode: 200, message: "success" });
+    const { data } = req.body;
+    const { user }: any = req.query;
+    const token: any = req.headers.authorization?.split(" ")[1];
+
+    jwt.verify(token, process.env.NEXTAUTH_SECRET || "", async (err: any, decoded: any) => {
+      if (decoded && decoded.role === "admin") {
+        await updateData("users", user[1], data, (result: boolean) => {
+          if (result) {
+            res.status(200).json({ status: true, statusCode: 200, message: "success" });
+          } else {
+            res.status(400).json({ status: false, statusCode: 400, message: "failed" });
+          }
+        });
       } else {
-        res.status(400).json({ status: false, statusCode: 400, message: "failed" });
+        res.status(403).json({ status: false, statusCode: 403, message: "Access denied" });
       }
     });
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res.status(200).json({ status: true, statusCode: 200, message: "success" });
+    const token: any = req.headers.authorization?.split(" ")[1];
+
+    jwt.verify(token, process.env.NEXTAUTH_SECRET || "", async (err: any, decoded: any) => {
+      if (decoded && decoded.role === "admin") {
+        await deleteData("users", user[1], (result: boolean) => {
+          if (result) {
+            res.status(200).json({ status: true, statusCode: 200, message: "success" });
+          } else {
+            res.status(400).json({ status: false, statusCode: 400, message: "failed" });
+          }
+        });
       } else {
-        res.status(400).json({ status: false, statusCode: 400, message: "failed" });
+        res.status(403).json({ status: false, statusCode: 403, message: "Access denied" });
       }
     });
   }
